requirements.txt
pi3d~=2.41
numpy>=1.19.5
Pillow~=8.1.0
geopy~=2.1.0
sqlalchemy~=1.3.23
python-dateutil~=2.8.1